refactor(test): split webhook tests by type
